Realocando métodos de Admin usuário e inserindo criação de senha na controller
