+++ Api/Controllers/AdmController.cs
@@ -2,6 +2,11 @@
 using CrossCutting.Configuration.Authorization;
 using Domain.Commands.v1.Adm.AlteraStatusUser;
 using Domain.Commands.v1.Login;
+using Domain.Commands.v1.Usuarios.AtualizarUsuario;
+using Domain.Commands.v1.Usuarios.BuscarUsuarioPorId;
+using Domain.Commands.v1.Usuarios.CriarUsuario;
+using Domain.Commands.v1.Usuarios.ListarUsuarios;
+using Domain.Commands.v1.Usuarios.RemoverUsuario;
 using Infrastructure.Data;
 using Infrastructure.Data.Models.Adm.AlteraStatusUser;
 using MediatR;
@@ -31,6 +36,76 @@ namespace Api.Controllers
         {
             var result = await _mediator.Send(command);
             return Ok(result);
+        }
+
+        [HttpGet]
+        [Route(template: "BuscarUsuarios")]
+        [Authorize(Policy = PoliticasDeAcesso.Usuario)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public async Task<IActionResult> BuscarTodos()
+        {
+            var query = new ListarUsuariosCommand();
+            var usuarios = await _mediator.Send(query);
+
+            return Ok(usuarios);
+        }
+
+        [HttpPost]
+        [Route(template: "Criar")]
+        [ProducesResponseType(typeof(CriarUsuarioCommandResponse), StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public async Task<IActionResult> Criar([FromBody] CriarUsuarioCommand command)
+        {
+            var response = await _mediator.Send(command);
+            return Created(string.Empty, response);
+        }
+
+        [HttpGet]
+        [Route(template: "BuscarPorId/{id}")]
+        [Authorize(Policy = PoliticasDeAcesso.Usuario)]
+        [ProducesResponseType(typeof(BuscarUsuarioPorIdCommandResponse), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public async Task<IActionResult> BuscarPorId(Guid id)
+        {
+            var query = new BuscarUsuarioPorIdCommand(id);
+            var usuario = await _mediator.Send(query);
+
+            return Ok(usuario);
+        }
+
+        [HttpPut]
+        [Route(template: "Atualizar")]
+        [Authorize(Policy = PoliticasDeAcesso.Usuario)]
+        [ProducesResponseType(typeof(AtualizarUsuarioCommandResponse), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public async Task<IActionResult> Atualizar([FromBody] AtualizarUsuarioCommand command)
+        {
+            return Ok(await _mediator.Send(command));
+        }
+
+        [HttpDelete]
+        [Route(template: "Remover")]
+        [Authorize(Policy = PoliticasDeAcesso.Usuario)]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public async Task<IActionResult> Remover(Guid id)
+        {
+            var command = new RemoverUsuarioCommand(id);
+            await _mediator.Send(command);
+
+            return NoContent();
         }
     }
 }
