+++ Api/Controllers/UsuarioController.cs
@@ -1,6 +1,7 @@
 ﻿using CrossCutting.Configuration.Authorization;
 using Domain.Commands.v1.Usuarios.AtualizarUsuario;
 using Domain.Commands.v1.Usuarios.BuscarUsuarioPorId;
+using Domain.Commands.v1.Usuarios.CriarSenha;
 using Domain.Commands.v1.Usuarios.CriarUsuario;
 using Domain.Commands.v1.Usuarios.ListarUsuarios;
 using Domain.Commands.v1.Usuarios.RemoverUsuario;
@@ -22,69 +23,18 @@ namespace Api.Controllers
             _mediator = mediator;
         }
 
-        [HttpGet]
-        [Authorize(Policy = PoliticasDeAcesso.Usuario)]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-        [ProducesResponseType(StatusCodes.Status403Forbidden)]
-        public async Task<IActionResult> BuscarTodos()
-        {
-            var query = new ListarUsuariosCommand();
-            var usuarios = await _mediator.Send(query);
+        
 
-            return Ok(usuarios);
-        }
-
+        [Route(template: "CriarSenha/{email}")]
         [HttpPost]
-        [ProducesResponseType(typeof(CriarUsuarioCommandResponse), StatusCodes.Status201Created)]
+        [AllowAnonymous]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-        [ProducesResponseType(StatusCodes.Status403Forbidden)]
-        public async Task<IActionResult> Criar([FromBody] CriarUsuarioCommand command)
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public async Task<IActionResult> CriarSenha([FromBody] CriarSenhaCommand command)
         {
             var response = await _mediator.Send(command);
             return Created(string.Empty, response);
-        }
-
-        [HttpGet("{id}")]
-        [Authorize(Policy = PoliticasDeAcesso.Usuario)]
-        [ProducesResponseType(typeof(BuscarUsuarioPorIdCommandResponse), StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-        [ProducesResponseType(StatusCodes.Status403Forbidden)]
-        public async Task<IActionResult> BuscarPorId(Guid id)
-        {
-            var query = new BuscarUsuarioPorIdCommand(id);
-            var usuario = await _mediator.Send(query);
-
-            return Ok(usuario);
-        }
-
-        [HttpPut("{id}")]
-        [Authorize(Policy = PoliticasDeAcesso.Usuario)]
-        [ProducesResponseType(typeof(AtualizarUsuarioCommandResponse), StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-        [ProducesResponseType(StatusCodes.Status403Forbidden)]
-        public async Task<IActionResult> Atualizar([FromBody] AtualizarUsuarioCommand command)
-        {
-            return Ok(await _mediator.Send(command));
-        }
-
-        [HttpDelete("{id}")]
-        [Authorize(Policy = PoliticasDeAcesso.Usuario)]
-        [ProducesResponseType(StatusCodes.Status204NoContent)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
-        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-        [ProducesResponseType(StatusCodes.Status403Forbidden)]
-        public async Task<IActionResult> Remover(Guid id)
-        {
-            var command = new RemoverUsuarioCommand(id);
-            await _mediator.Send(command);
-
-            return NoContent();
         }
     }
 }
